Change the initialization code for better performance
Opening a zip file for reading was causing the whole metadata to be dumped
to a db. This is a very slow process for archives with a lot of files and
it doesn't seem to bring any major performance enhancement in accessing
the single files within the zip. Switched to a lazy cache implementation.